Refactor archives
- Ruby code split
- Use webpack-manifest-plugin
