+++ public/plugin/lokka-archives/.gitignore
@@ -1,0 +1,1 @@
+assets/*.js
