+++ public/plugin/lokka-archives/assets/manifest.json
@@ -1,0 +1,3 @@
+{
+  "index.js": "index-a5303c782bbdfb1fc2fb.js"
+}
