+++ public/plugin/lokka-archives/lib/lokka/archives.rb
@@ -1,15 +1,17 @@
 # frozen_string_literal: true
 
+require_relative 'archives/chart_query_generator'
+require_relative 'archives/aggregator'
+
 module Lokka
   module Archives
     def self.registered(app)
       app.get '/archives.json' do
         posts = Post.published.joins(:category)
-        month_posts = MonthPosts.generate(posts)
 
         cache_control :public, :must_revalidate, max_age: 5.minutes
         content_type :json
-        month_posts.to_json
+        month_posts(posts).to_json
       end
 
       app.get '/archives/years.json' do
@@ -31,10 +33,9 @@ module Lokka
         posts = Post.published.joins(:category).where(
           created_at: (Time.new(year)..Time.new(year).end_of_year)
         )
-        month_posts = MonthPosts.generate(posts)
 
         content_type :json
-        month_posts.to_json
+        month_posts(posts).to_json
       end
 
       app.get '/archives' do
@@ -60,185 +61,34 @@ module Lokka
     end
 
     def categories
-      MonthPosts.categories.map(&:title)
+      Archives::Aggregator.categories.map(&:title)
+    end
+
+    def month_posts(posts)
+      Archives::Aggregator.generate(posts)
     end
 
     def chart
-      sql_chart || ruby_chart
+      Archives::ChartQueryGenerator.run
     end
 
-    private
-
-    def sql_chart
-      query = case Lokka.database_config.to_s
-              when /mysql/
-                mysql_query
-              when /postgres/
-                postgres_query
-              when /sqlite/
-                sqlite_query
-              else
-                nil
-              end
-      return nil if query.nil?
-
-      result = ActiveRecord::Base.connection.select_all(query)
-      grouped = result.map(&:to_h).group_by {|record| record['year'] }
-      grouped.each_with_object([]) {|(year, record), outer|
-        outer << record.each_with_object({ year: year }) {|item, inner|
-          _, category, count = item.values
-          inner[category] = count
-        }
-      }
+    def archives_assets_path
+      '/plugin/lokka-archives/assets'
     end
 
-    def postgres_query
-      <<~SQL
-        SELECT
-          TO_CHAR(entries.created_at, 'YYYY') AS year,
-          categories.title as category,
-          COUNT(1) AS count
-        FROM entries
-        INNER JOIN categories ON categories.id = entries.category_id
-        WHERE entries.draft = FALSE AND entries.type = 'Post'
-        GROUP BY year, category
-        ORDER BY year
-      SQL
+    def manifest
+      @manifest ||= begin
+                      file_path = File.join(Lokka.root, 'public', archives_assets_path, 'manifest.json')
+                      content = File.open(file_path).read
+                      manifest = JSON.parse(content)
+                    end
     end
 
-    def mysql_query
-      <<~SQL
-        SELECT
-          YEAR(entries.created_at) AS year,
-          categories.title as category,
-          COUNT(1) AS count
-        FROM entries
-        INNER JOIN categories ON categories.id = entries.category_id
-        WHERE entries.draft = FALSE AND entries.type = 'Post'
-        GROUP BY year, category
-        ORDER BY year
-      SQL
-    end
-
-    def sqlite_query
-      <<~SQL
-        SELECT
-          strftime('%Y', entries.created_at) AS year,
-          categories.title as category,
-          COUNT(1) AS count
-        FROM entries
-        INNER JOIN categories ON categories.id = entries.category_id
-        WHERE entries.draft = 'f' AND entries.type = 'Post'
-        GROUP BY year, category
-        ORDER BY year
-      SQL
-    end
-
-    def ruby_chart
-      posts = Post.unscoped.published.select(%i[id category_id created_at]).includes(:category)
-      grouped = posts.group_by {|post| [post.created_at.year, post.category&.title] }
-      group_by_year = grouped.group_by {|(year, category), entries| year }
-      group_by_year.each_with_object([]) {|(year, records), outer|
-        outer << records.each_with_object({ year: year }) {|((_, category), entries), inner|
-          inner[category] = entries.length
-        }
-      }.reverse
-    end
-  end
-
-  module AddDateTimeMethodsToPost
-    refine Post do
-      def year
-        created_at.year.to_s.rjust(4, '0')
-      end
-
-      def monthnum
-        created_at.month.to_s.rjust(2, '0')
-      end
-
-      def month
-        created_at.month.to_s.rjust(2, '0')
-      end
-
-      def day
-        created_at.day.to_s.rjust(2, '0')
-      end
-
-      def hour
-        created_at.hour.to_s.rjust(2, '0')
-      end
-
-      def minute
-        created_at.min.to_s.rjust(2, '0')
-      end
-
-      def second
-        created_at.sec.to_s.rjust(2, '0')
-      end
-
-      def post_id
-        id.to_s
-      end
-
-      def postname
-        slug || id.to_s
-      end
-
-      def clever_link
-        if permalink_format
-          params = permalink_keys.each_with_object({}) {|key, hash| hash[key] = send(key) }
-          permalink_helper.custom_permalink_path(params)
-        else
-          slug
-        end
-      end
-
-      private
-
-      def permalink_helper
-        Lokka::PermalinkHelper
-      end
-
-      def permalink_format
-        @permalink_format ||= permalink_helper.custom_permalink_format if permalink_helper.custom_permalink?
-      end
-
-      def permalink_keys
-        @permalink_keys ||=
-          permalink_format.map {|item| item.sub(%r{(?:%(.+?)%)?/?}, '\1') }.
-            delete_if(&:blank?).map(&:to_sym)
-      end
-    end
-  end
-
-  class MonthPosts
-    using AddDateTimeMethodsToPost
-
-    class << self
-      def categories
-        @categories ||= Category.find(
-          Category.joins(:entries).where(entries: Post.published).
-          group(:id).order(count_entries_id: :desc).count(:'entries.id').
-          keys
-        )
-      end
-
-      def generate(posts)
-        posts.group_by {|post| post.created_at.beginning_of_month.to_formatted_s(:db) }.
-          each_with_object({}) do |(month, month_posts), object|
-            object[month] ||= []
-            month_posts.each do |post|
-              category = categories.find {|c| c.id == post.category_id }
-              object[month] << {
-                id: post.id,
-                category: { id: category[:id], title: category[:title], slug: category[:slug] },
-                title: post.title,
-                link: post.clever_link,
-                created_at: post.created_at
-              }
-            end
-            object
-          end
+    def archives_javascript_path(file_name)
+      if Lokka.production?
+        "#{archives_assets_path}/#{manifest[file_name]}"
+      else
+        "/plugin/lokka-archives/build/#{file_name}"
       end
     end
   end
